requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
cattrs==23.1.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.55.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-phone-field==1.8.1
django-phonenumber-field==7.2.0
django-schema-graph==3.1.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
kombu==5.3.2
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.23
Pillow==10.1.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.6.0
zope.interface==6.1
